backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.18.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
stripe==8.0.0
sqlalchemy==2.0.25
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
httpx==0.26.0
pytest==8.0.0
pytest-asyncio==0.23.3
